GradebookREADME.md
The main objective of the Gradebook Management System is to automate the process of calculating students' The Gradebook Management System effectively automates the grading process, providing educators with an efficient tool to manage students' academic performance. Its user-friendly interface and accurate calculations make it a valuable asset for educators in evaluating students' grades.